\name{DocTopics}
\alias{DocTopics}
\title{Find the topic estimated that a document is most likely to be from.}
\usage{
  DocTopics(obj, TopicLabels = NULL)
}
\arguments{
  \item{obj}{an \code{ExpAgendaOut} class object created by
  \code{\link{ExpAgendaVonmon}}.}

  \item{TopicLabels}{character vector. Labels for each
  topic. Should be in the same order as the topics
  generated by \code{\link{TopicSummary}}.}
}
\value{
  A data frame with three columns: \code{ID}, \code{Names},
  \code{Topic}.
}
\description{
  Find the topic estimated that a document is most likely
  to be from.
}

