\name{DocTopics}
\alias{DocTopics}
\title{Find the topic estimated that a document is most likely to be from.}
\usage{
  DocTopics(obj, TopicLabels = "auto", StemLabel = 2)
}
\arguments{
  \item{obj}{an \code{ExpAgendaOut} class object created by
  \code{\link{ExpAgendaVonmon}}.}

  \item{TopicLabels}{character vector. Labels for each
  topic. Should be in the same order as the topics
  generated by \code{\link{TopicSummary}}. If
  \code{TopicLabels = "auto"} then the top topic stems are
  used.}

  \item{StemLabels}{numeric. The number of stems to use for
  \code{auto} generated topic labels.}
}
\value{
  A data frame with three columns: \code{ID}, \code{Names},
  \code{Topic}.
}
\description{
  Find the topic estimated that a document is most likely
  to be from.
}

