\name{DocTopics}
\alias{DocTopics}
\title{Find the topic estimated that a document is most likely to be from.}
\usage{
  DocTopics(obj)
}
\arguments{
  \item{obj}{an \code{ExpAgendaOut} class object created by
  \code{\link{ExpAgendaVonmon}}.}
}
\description{
  Find the topic estimated that a document is most likely
  to be from.
}

